+++ tests/test_ops.py
@@ -191,6 +191,43 @@ def test_divide_scenario(test_df_year, append):
         assert_iamframe_equal(exp, obs)
 
 
+@pytest.mark.parametrize("append", (False, True))
+def test_apply_variable(plot_stackplot_df, append):
+    """Verify that in-dataframe apply works on the default `variable` axis"""
+
+    def custom_func(a, b, c, d, e):
+        return a / c + b / d + e
+
+    args = ["Emissions|CO2|Tar", "Emissions|CO2|Cars", "Emissions|CO2|LUC"]
+    kwds = {"d": "Emissions|CO2|Agg", "e": 5}
+    exp = IamDataFrame(
+        pd.DataFrame(
+            [
+                0.3 / (-0.3) + 1.6 / 0.5 + 5,
+                0.35 / (-0.6) + 3.8 / (-0.1) + 5,
+                0.35 / (-1.2) + 3.0 / (-0.5) + 5,
+                0.33 / (-1.0) + 2.5 / (-0.7) + 5,
+            ],
+            index=[2005, 2010, 2015, 2020],
+        ).T,
+        model="IMG",
+        scenario="a_scen",
+        region="World",
+        variable="new variable",
+        unit="Mt CO2/yr",
+    )
+
+    if append:
+        obs = plot_stackplot_df.copy()
+        obs.apply(custom_func, name="new variable", append=True, args=args, **kwds)
+        assert_iamframe_equal(plot_stackplot_df.append(exp), obs)
+    else:
+        obs = plot_stackplot_df.apply(
+            custom_func, name="new variable", args=args, **kwds
+        )
+        assert_iamframe_equal(exp, obs)
+
+
 def test_ops_unknown_axis(test_df_year):
     """Using an unknown axis raises an error"""
     with pytest.raises(ValueError, match="Unknown axis: foo"):